requirements.txt
fastapi[standard]
pyjwt
cryptography
